+++ mobile/android/geckoview/src/main/java/org/mozilla/geckoview/doc-files/CHANGELOG.md
@@ -21,12 +21,15 @@ exclude: true
   ([bug 1674428]({{bugzilla}}1674428))
 - Added [`GeckoSession.PromptDelegate.onRequestCertificate`][138.5] to support client authentication certificates.
   ([bug 1813930]({{bugzilla}}1813930))
+- Added [`Loader#textDirectiveUserActivation`][138.6] to set text directive user activation.
+  ([bug 1912095]({{bugzilla}}1912095))
 
 [138.1]: {{javadoc_uri}}/GeckoSession.Loader.html#originalInput(java.lang.String)
 [138.2]: {{javadoc_uri}}/GeckoSession.PromptDelegate.FilePrompt.Type.html#FOLDER
 [138.3]: {{javadoc_uri}}/GeckoSession.PromptDelegate.html#onFolderUploadPrompt(org.mozilla.geckoview.GeckoSession,org.mozilla.geckoview.GeckoSession.PromptDelegate.FolderUploadPrompt)
 [138.4]: {{javadoc_uri}}/GeckoSession.PromptDelegate.FolderUploadPrompt.html
 [138.5]: {{javadoc_uri}}/GeckoSession.PromptDelegate.html#onRequestCertificate(org.mozilla.geckoview.GeckoSession,org.mozilla.geckoview.GeckoSession.PromptDelegate.CertificateRequest)
+[138.6]: {{javadoc_uri}}/GeckoSession.Loader.html#textDirectiveUserActivation(boolean)
 
 ## v137
 - ⚠️ [`GeckoSession.requestAnalysis`][118.4], [`GeckoSession.requestCreateAnalysis`][122.2], [`GeckoSession.requestAnalysisStatus`][137.1], [`GeckoSession.sendPlacementAttributionEvent`][123.3], [`GeckoSession.pollForAnalysisCompleted`][137.2], [`GeckoSession.sendClickAttributionEvent`][121.4], [`GeckoSession.sendImpressionAttributionEvent`][121.5], [`GeckoSession.sendPlacementAttributionEvent`][123.3], [`GeckoSession.requestRecommendations`][118.5], [`GeckoSession.reportBackInStock`][122.1], `AnalysisStatusResponse`, [`ReviewAnalysis`][120.2] and [`Recommendation`][120.3] are deprecated, and it will be deleted in version 139 see https://bugzilla.mozilla.org/show_bug.cgi?id=1941470.
@@ -1692,4 +1695,4 @@ to allow adding gecko profiler markers.
 [65.24]: {{javadoc_uri}}/CrashReporter.html#sendCrashReport(android.content.Context,android.os.Bundle,java.lang.String)
 [65.25]: {{javadoc_uri}}/GeckoResult.html
 
-[api-version]: c1976a2f8a7a3986f07426c8daccfc0bd0dcc746
+[api-version]: e70347dcd46ef50841e7ae06714dbc69f7036e85